# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "INEED API DOC",
		"description": "OpenAPI Specification for INEED API"
	},
	"schemes": [
		"http"
	],
	"host": "localhost:4002",
	"basePath": "/api",
...
	"paths": {
...
		"/user": {
			"get": {
				"tags": [
					"User"
				],
				"summary": "Get user profile",
				"description": "Get user profile functionality for users.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"put": {
				"tags": [
					"User"
				],
				"summary": "Update user",
				"description": "Functionality for user to update profile data.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"in": "body",
						"name": "Update user",
						"description": "Update user request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"firstName": {
									"type": "string"
								},
								"lastName": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/user/phone": {
			"put": {
				"tags": [
					"User"
				],
				"summary": "Update user phone",
				"description": "Functionality for user to update phone number.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"in": "body",
						"name": "Update user phone",
						"description": "Update user phone request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"phoneNumber": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/user/verifyPhone": {
			"post": {
				"tags": [
					"User"
				],
				"summary": "Verify phone number",
				"description": "Functionality for user to verify phone number.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"in": "body",
						"name": "Verify phone number",
						"description": "Verify phone number request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"phoneNumber": {
									"type": "string"
								},
								"code": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
...
		"/conversation/user": {
			"get": {
				"tags": [
					"Conversation"
				],
				"summary": "Functionality for APP Users, or APP Guests to get message list.",
				"description": "Functionality for APP Users, or APP Guests to get message list.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/AuthorizationOptional"
					},
					{
						"name": "deviceId",
						"in": "header",
						"required": false,
						"description": "Device Id of the app guest",
						"type": "string"
					},
					{
						"name": "language",
						"in": "header",
						"required": true,
						"description": "Language",
						"type": "string"
					},
					{
						"name": "skip",
						"in": "query",
						"required": true,
						"description": "Message count to skip",
						"type": "number"
					},
					{
						"$ref": "#/parameters/limit"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
...
		"/partner": {
			"post": {
				"tags": [
					"Partner"
				],
				"summary": "Add partner",
				"description": "Functionality to create new partner.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"in": "body",
						"name": "Add partner",
						"description": "Add partner request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"vatid": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"email": {
									"type": "string"
								},
								"message": {
									"type": "string"
								},
								"phoneNumber": {
									"type": "string"
								},
								"contactperson": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/partner/request": {
			"post": {
				"tags": [
					"Partner"
				],
				"summary": "Add partner",
				"description": "Functionality to create new partner.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"in": "body",
						"name": "Add partner",
						"description": "Add partner request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"vatid": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"email": {
									"type": "string"
								},
								"message": {
									"type": "string"
								},
								"phoneNumber": {
									"type": "string"
								},
								"contactperson": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/partner/selectList": {
			"get": {
				"tags": [
					"Partner"
				],
				"summary": "Functionality for APP Users, or APP Guests to get message badge.",
				"description": "Functionality for APP Users, or APP Guests to get message badge.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"in": "body",
						"name": "Add partner",
						"description": "Add partner request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"vatid": {
									"type": "string"
								},
								"name": {
									"type": "string"
								},
								"email": {
									"type": "string"
								},
								"message": {
									"type": "string"
								},
								"phoneNumber": {
									"type": "string"
								},
								"contactperson": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/partner/activatePartner": {
			"put": {
				"tags": [
					"Partner"
				],
				"summary": "Functionality for Activate or Deactivate Partners using Admin permissions",
				"description": "Functionality for Activate or Deactivate Partners using Admin permissions",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"in": "body",
						"name": "Activate partner",
						"description": "Activate partner request body",
						"required": "true",
						"schema": {
							"type": "object",
							"properties": {
								"partnerId": {
									"type": "string"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/partner/list": {
			"get": {
				"tags": [
					"Partner"
				],
				"summary": "Functionality for getting partners lists.",
				"description": "Functionality for getting partners lists.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"$ref": "#/parameters/pageNo"
					},
					{
						"$ref": "#/parameters/limit"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
		"/partner/details": {
			"get": {
				"tags": [
					"Partner"
				],
				"summary": "Functionality for getting partner details.",
				"description": "Functionality for getting partner details.",
				"consumes": "application/json",
				"produces": "application/json",
				"parameters": [
					{
						"$ref": "#/parameters/Authorization"
					},
					{
						"name": "partnerId",
						"in": "query",
						"required": true,
						"description": "Partner identifier",
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Successful response object"
					},
					"400": {
						"description": "Bad request, check keys and values"
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		}
	},
	"parameters": {
		"Authorization": {
			"name": "Authorization",
			"in": "header",
			"required": true,
			"description": "Bearer token",
			"type": "string"
		},
		"AuthorizationOptional": {
			"name": "Authorization",
			"in": "header",
			"required": false,
			"description": "Bearer token",
			"type": "string"
		},
...
		"pageNo": {
			"name": "pageNo",
			"in": "query",
			"required": true,
			"description": "Page number",
			"type": "number"
		},
		"limit": {
			"name": "limit",
			"in": "query",
			"required": true,
			"description": "Limit to get",
			"type": "number"
		},
...
	}
}